Minor bugs
git-svn-id: https://apertium.svn.sourceforge.net/svnroot/apertium@1290 72bbbca6-d526-0410-a7d9-f06f51895060

+++ apertium/apertium/LexTorData.C
@@ -413,9 +413,9 @@ LexTorData::reduce(const string& s) {
     }
   }
 
-  unsigned p=str.find(">");
-  unsigned i=0;
-  if (p==string::npos) { //s could correspond to an unknown word
+  unsigned int p=str.find(">");
+  unsigned int i=0;
+  if (p==static_cast<unsigned int>(string::npos)) { //s could correspond to an unknown word
     p=str.length();
     if ((str.length()>0) && (str[0]=='*'))
       i=1; // to remove the star (unknown word mark)
